Move tests to separate test_dlpack.py
